# MW-946: Aggregated common operations on group starts