--- a/application/models/services/Proxy/ProxyExpressionManager.php
+++ b/application/models/services/Proxy/ProxyExpressionManager.php
@@ -64,13 +64,14 @@ class ProxyExpressionManager
                 $aGroup['gid'],
                 $survey->anonymized != 'Y',
                 $surveyId,
-                false,
+                true,
                 true
             );
         }
         LimeExpressionManager::AggregateGroupStart(
             $survey->anonymized != 'Y',
-            $surveyId
+            $surveyId,
+            true
         );
         foreach ($aGrouplist as $aGroup) {
             LimeExpressionManager::FinishProcessingGroup();