--- a/application/models/services/Proxy/ProxyExpressionManager.php
+++ b/application/models/services/Proxy/ProxyExpressionManager.php
@@ -64,6 +64,7 @@ class ProxyExpressionManager
                 $aGroup['gid'],
                 $survey->anonymized != 'Y',
                 $surveyId,
+                false,
                 true
             );
         }