This is a performance improvement for MW-946, the idea is to improve the speed of the `reset` method by aggregating the survey-level operations that were previously done for each group and thus performing the repetitive operations only once.

Measurements:

Before:

Language Settings took 0.010984897613525 seconds
General Settings took 0.0005490779876709 seconds
URL Params took 9.5367431640625E-7 seconds
Proxy Reset took 12.65425491333 seconds
Template Config took 0.0011730194091797 seconds
Database::actionUpdateSurveyLocaleSettings took 12.67779302597

After:

Language Settings took 0.012439012527466 seconds
General Settings took 0.00066709518432617 seconds
URL Params took 9.5367431640625E-7 seconds
Proxy Reset took 3.8038971424103 seconds
Template Config took 0.001168966293335 seconds
Database::actionUpdateSurveyLocaleSettings took 3.828782081604 seconds