@Shnoulle 2024-02-14
ProxyExpressionManager->reset seems to replace/reset the ExpressionManager session when something is updated in Survey.

We need to be sure it was done when

- Admin add/remove/rename a question in LimeSurvey
- Admin user click on preview or survey logic file
- Somple user start with newtest=Y (or empty session)

OK with this ?

If yes 

1. why not `LimeExpressionManager::AggregateGroupStart` before [`foreach ($aGrouplist as $aGroup) {` (line](https://github.com/LimeSurvey/LimeSurvey/blob/38093981a2c05e9ec2cfe5f7e305330870eadd41/application/models/services/Proxy/ProxyExpressionManager.php#L62)  ?
2. Use `AggregateGroupStart($survey->anonymized != 'Y', $surveyId, true)` (force refresh) 


I don't find where ProxyExpressionManager->reset is call : you know here ?