@lajosarpad 2024-02-14
> ProxyExpressionManager->reset seems to replace/reset the ExpressionManager session when something is updated in Survey.
> 
> We need to be sure it was done when
> 
> * Admin add/remove/rename a question in LimeSurvey
> * Admin user click on preview or survey logic file
> * Simple user start with newtest=Y (or empty session)
> 
> OK with this ?
> 
> If yes
> 
> 1. why not `LimeExpressionManager::AggregateGroupStart` before [`foreach ($aGrouplist as $aGroup) {` (line](https://github.com/LimeSurvey/LimeSurvey/blob/38093981a2c05e9ec2cfe5f7e305330870eadd41/application/models/services/Proxy/ProxyExpressionManager.php#L62)  ?
> 2. Use `AggregateGroupStart($survey->anonymized != 'Y', $surveyId, true)` (force refresh)
> 
> I don't find where ProxyExpressionManager->reset is call : you know here ?
> 
> (i do some test when found some times)

@Shnoulle `ProxyExpressionManager->reset` is being called at `SurveyAggregateService::update` at the section of:

```
        $this->proxyExpressionManager
            ->reset($surveyId);
```

This is being triggered for example from `Database::actionUpdateSurveyLocalSettings`.

"We need to be sure it was done when

Admin add/remove/rename a question in LimeSurvey
Admin user click on preview or survey logic file
Simple user start with newtest=Y (or empty session)
OK with this ?"

I do not have enough context to answer this question. But if we take a look at the code before the patch, we will see that:

- we have a foreach inside application/models/services/Proxy/ProxyExpressionManager.php
- which calls `LimeExpressionManager::StartProcessingGroup` for each group
- and `StartProcecssingGroup` calls the methods that are being called in `AggregateGroupStart`
- these methods in the earlier version are being called after the main group handling of `StartProcessingGroup` has been done

So, these are common costs of the groups and the idea was to avoid repetitively doing these costly operations for each groups, and, instead, to do it once. The reason why we have it after the first foreach is that it seemed to me, given the position of these methods calls inside `StartProcessingGroup` that it is more appropriate to call them after the groups have started. But I'm not sure about this and if it is more appropriate to do it before the groups are initialized, then we can move the call for `AggregateGroupStart` prior the foreach. My thinking was that maybe $fieldMap changes during the group initialization which makes it necessary to call `setVariableAndTokenMappingsForExpressionManager` but I could be wrong about this.

The reason as of why force refresh was left false was that it was left force before the patch and I presumed it's okay that way:

![image](https://github.com/LimeSurvey/LimeSurvey/assets/2805488/bb58fb99-80a1-4bb1-af30-55ff5c872b58)

What the patch does in short:

- it makes the calls for `setVariableAndTokenMappingsForExpressionManager`, `ProcessAllNeededRelevance` and `_CreateSubQLevelRelevanceAndValidationEqns` only if we want to (this is why the new parameter of `$postponeAggregation` was created)
- we do not want to call these methods for all groups in `ProxyExpressionManager::reset`, we postpone them to make sure we call them exactly once
- and this is why `AggregateGroupStart` was implemented

The tests were successful (I tested by changing the survey settings) and a process that was taking 13-17s prior the patch is now taking ~ 4s for the same survey. So, I think the idea is good, but we need your and ptelu's take on this to make sure all the details are correct. And, if some details need improvement, we will make sure they are being handled as instructed.

Thanks for the review and the questions.