@Shnoulle 2024-02-15
> So, these are common costs of the groups and the idea was to avoid repetitively doing these costly operations for each groups, and, instead, to do it once.

Yes : i totally understand this , and clearly : doing it for each group is an issue :)


> given the position of these methods calls inside StartProcessingGroup that it is more appropriate to call them after the groups have started. But I'm not sure about this and if it is more appropriate to do it before the groups are initialized, then we can move the call for AggregateGroupStart prior the foreach.

OK, i test the 2 position, it's seemslogical for me to «update the _VariableAndTokenMappingsForExpressionManager_ for all survey, use new _VariableAndTokenMappingsForExpressionManager_ for all group» :) 

I try to understand better what happen adding some print_r and log.

> The reason as of why force refresh was left false was that it was left false before the patch and I presumed it's okay that way:

OK, surely resetted before (set to empty session)