@lajosarpad 2024-02-15
> OK, i test the 2 position, it's seemslogical for me to «update the VariableAndTokenMappingsForExpressionManager for all survey, use new VariableAndTokenMappingsForExpressionManager for all group» :)

Yes, the case you are making is quite plausible. If you can confirm that it is more appropriate to put the call for it before the loop, then I will make the change accordingly. As about force refresh, I have implemented the change you have suggested.