@Shnoulle 2024-02-15
> As about force refresh, I have implemented the change you have suggested.

No ? Maybe it's not needed always ???