@lajosarpad 2024-02-15
> > As about force refresh, I have implemented the change you have suggested.
> 
> No ? Maybe it's not needed always ???

Should I set it back to false, like it was earlier, before https://github.com/LimeSurvey/LimeSurvey/pull/3734/commits/d3c5ed50907858b4d936f2d626cc07faa2b6eab4 ?

The code before my change used false as force refresh and I preserved the same behavior. I think I misunderstood your point when I changed it to true. I can change it back to false.