@Shnoulle 2024-02-15
> The code before my change used false as force refresh and I preserved the same behavior. I think I misunderstood your point when I changed it to true. I can change it back to false.

It's more i totally unsure on this point. Maybe it's already resetted before, if we reset a second time : we do whole StatrSurvey a second time …