@Shnoulle 2024-02-15
Some test :  hard to find when this was used … and since when testing : it resetted by session : i think you can remove the reset in `LimeExpressionManager::StartProcessingGroup`

It must be done only one time (why not with `LimeExpressionManager::AggregateGroupStart` ;))

Else: i try to find where it's used … i try updating question, preview, logic file …
 
**I feel incompetent to give an informed opinion here.**

Another issue find in logic file : we reset the setVariableAndTokenMappingsForExpressionManager for each group : here we really need to do it once (but sureyLogicFile not used a lot)